New Header to support iOS 13
As found in the documentation at https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/sending_notification_requests_to_apns

A new header with the name `apns-push-type` has to be sent for devices running iOS 13. The possible values are `alert` for normal notifications and `background` for `content-available` notifications.

While at it I fixed the `apns-priority` header value for `content-available` notifications which should be 5 or less according to the documentation, while still allowing for an arbitrary number to be specified if the user wants to.